refactor utils and fix small bug
