+++ Composer/scripts/l10n-extract.js
@@ -4,22 +4,9 @@
 const fs = require('fs');
 const path = require('path');
 
-const L10N_FIELDS = ['label', 'description', 'title', 'subtitle'];
+const { keep, transFn } = require('./l10nUtils');
 
-function keep(obj, keptFields) {
-  const out = {};
-  for (const key of Object.keys(obj)) {
-    if (keptFields.includes(key)) {
-      out[key] = obj[key];
-    }
-    if (typeof obj[key] === 'object') {
-      const value = keep(obj[key], keptFields);
-      if (Object.keys(value).length === 0) continue;
-      out[key] = value;
-    }
-  }
-  return out;
-}
+const L10N_FIELDS = ['label', 'description', 'title', 'subtitle'];
 
 const inFiles = process.argv.slice(2);
 
@@ -27,8 +14,11 @@ for (const file of inFiles) {
   let schema;
   const baseName = path.basename(file);
   const outputDir = path.join(path.normalize(path.dirname(file)), 'locales', 'en-US');
-  // eslint-disable-next-line security/detect-non-literal-fs-filename
+  const outputDirTrans = path.join(path.normalize(path.dirname(file)), 'locales', 'en-US-pseudo');
+  /* eslint-disable security/detect-non-literal-fs-filename */
   fs.mkdirSync(outputDir, { recursive: true });
+  fs.mkdirSync(outputDirTrans, { recursive: true });
+  /* eslint-enable */
 
   try {
     // eslint-disable-next-line security/detect-non-literal-fs-filename
@@ -39,9 +29,11 @@ for (const file of inFiles) {
   }
 
   const output = keep(schema, L10N_FIELDS);
+  const outputTransformed = keep(schema, L10N_FIELDS, transFn);
   const outputFn = outputDir + path.sep + baseName;
 
   console.log('writing', outputFn);
-  // eslint-disable-next-line security/detect-non-literal-fs-filename
+  /* eslint-disable security/detect-non-literal-fs-filename */
   fs.writeFileSync(outputDir + path.sep + baseName, JSON.stringify(output, null, 4));
+  fs.writeFileSync(outputDirTrans + path.sep + baseName, JSON.stringify(outputTransformed, null, 4));
 }
